Added support to read and check transparency of blocks with id up to 4096. Replacement of IDs > 255 doesn't work yet.

+++ src/main/java/com/lishid/orebfuscator/obfuscation/BlockUpdate.java
@@ -94,7 +94,7 @@ public class BlockUpdate {
     }
 
     public static void AddBlockCheck(HashSet<Block> allBlocks, Block block) {
-        if ((OrebfuscatorConfig.isObfuscated((byte) block.getTypeId(), block.getWorld().getEnvironment() == Environment.NETHER) || OrebfuscatorConfig.isDarknessObfuscated((byte) block.getTypeId()))) {
+        if ((OrebfuscatorConfig.isObfuscated(block.getTypeId(), block.getWorld().getEnvironment() == Environment.NETHER) || OrebfuscatorConfig.isDarknessObfuscated(block.getTypeId()))) {
             allBlocks.add(block);
         }
     }
